Refactor AstEnvironment [0] move pure-logic functions up
Summary:
**About this stack**

I have an idea of how to represent the root-environment vs
overlay-environment types in a proper way, which if it works will
make the overlay code *much* easier to reason about by explicitly
distinguishing between overlay vs root environment operations.

A prerequisite for me to fully explore this idea is to extract as
much logic into ReadOnly as I can; this short stack is going to do
that by moving the export-resolving code into AstEnvironment.ReadOnly;
that's important because we don't want to duplicate it between the
overlay and root types.

**This diff**

Move a couple of pure functions that just transform data to the
top of AstEnvironment; this will ensure that they are available to
both ReadOnly and non-ReadOnly code after later refactors; it's
always good to move the plain old functions to the top of any ocaml
module with state because that makes it clear that they are
state-independent!

Reviewed By: grievejia

Differential Revision: D36889634

fbshipit-source-id: 8109417b1d51de2c70fbf7020739a8d9a07335f4