Remove `UpdateResult.read_only`
Summary:
I want to work on a refactor to the environment stack
in which we split the read-write interface into two
parts:
- one part (which for now I'm calling `Abstract` that owns
  mutation logic for cache management, but has only
  read-only access to the parent environment
- a second part (which for now I'm calling `Base` because
  eventually it will have a parallel but distinct sibling
  module called `Overlay` that has full read-write access
  all the way down the environment stack.

Now, it's been a bit silly to include `read_only` in
the `UpdateResult` type to date because, at least in the
current code, we have trivial access to the upstream
environment anyway and we can just use that to get a read_only.

There's been a kind of half-benefit in that by avoiding
direct access to `upstream_environment` we make it a bit less
likely that we'll accidentally call the non-readonly version,
but it's not any more type-safe.

But as soon as I do the `Abstract` refactor, it will become
trivially type-safe that all code in `Abstract` has
exclusively read-only access to the lower layers. As a result,
it makes sense to first strip out this `UpdateResult.read_only`
logic.

Reviewed By: grievejia

Differential Revision: D36902473

fbshipit-source-id: a27652906dd91da17b716ece4ac9de625a8ea1f6