Make each environment layer own its upstream.
Summary:
See the previous commit message - the TLDR is that in order to
make sure that push-style updates work correctly, we want to prevent
a (read-write) environment layer from being directly accessed if
a higher-level layer exists. We can accomplish this by always
having the higher-level layer's constructor function create the lower-level
layer and then not exposing the lower layer in any apis.

This prevents anyone from accidentally creating invalid states (for example
two AnnotatedGlobalEnvironments built off of the same AstEnvironment.t value,
which is never valid but was possible to do prior to this commit).

In order to accomplish this, we expose three types of constructor functions:
- create, used to make brand-new environments from scratch in production
- load, used to get an environment from saved state
- create_for_testing, used to create an environment backed by in-memory
  sources (which causes that environment to be non-updatable; ModuleTracker
  knows about this and fails with a good error message if you mess up)

Differential Revision: D36882667

fbshipit-source-id: 626fba70d1dc1fe4f2c976a944c8bb86ee7376d6