Move ModuleTracker.ReadOnly to top
Summary:
This is the first of two commits where I move the
read-write logic for ModuleTracker into a Base module
so that it is easy to isolate from the (different)
functionality that we will need in an Overlay-based
read-write environment.

This commit is trivial, it's just moving ReadOnly
to the top which is a prereq for wrapping together all
of the read-write functionality.

Reviewed By: grievejia

Differential Revision: D36921515

fbshipit-source-id: 5a2757f3a52ada403948c66d16285d9f5f26803d