Have AstEnvironment fully own the module tracker
Summary:
The root cause of a lot of our problems with testing APIs -
which also makes production codepaths harder to understand - is that
instead of creating the entire environment stack at once we create
some of the bottom layers first and then pass them into higher layers.

This dependency injection pattern might be good for its flexibility,
except that there's a catch: the control flow in push updates means
that we need a higher-level environment layer to *completely* own
lower levels, because it creates illegal states if the lower levels
are ever updated without updating the higher levels.

This is something we can fix by having appropriate creation apis
at all layers so that they can construct their own (hidden) inner
layers, which guarantees that the update rules are never violated.

To do this, we need to have three creation apis:
- create, used for all from-scratch environments in production
- load, used to load from a saved state
- create_for_testing, used to create an environment that is backed
  by in-memory sources (and cannot be updated).

This commit updates AstEnvironment to expose these three apis,
in the next commit I'll update the higher-level environments to
do the same.

Differential Revision: D36881932

fbshipit-source-id: dfb58aa4361157bfbb5bdce1e7a88972504485c2