Standardize TypeEnvironment [0] ReadOnly to top
Summary:
Before I start working on overlays, I'd like to make TypeEnvironment
at least a little bit less "wierd" than it is now by standardizing
the API to look more like a normal EnvironmentTable-based environment.

I haven't decided yet whether I want to actually try to
make it use EnvironmentTable - which I'm pretty sure is possible
to do and would simiplify some of the dependency tracking logic
quite a bit - but at the very least I want it to be factored in
the same general way so that we can have ReadOnly, Base, and
Overlay versions of the implementation because that's a hard-blocker
on overlays.

This first commit moves ReadOnly to the top so that I can push the
read-write logic into a Base module where it will be isolated from
Overlay logic once we add that.

Reviewed By: grievejia

Differential Revision: D36921757

fbshipit-source-id: b032790917534f0cccd0e071616cd0fde8f222c6