Identify expressions with TypeIsAny coverage gaps
Summary:
### Context:
Pyre catches errors in code only when the following are true:
Users have added types to their code
Libraries have types
The Python feature used is supported by Pyre

### Problem:
When this is not true, there is a coverage gap, which signals that a runtime crash is possible. Hence, we want to identify expressions that contain coverage gaps.

### Solution:
To do this, we add a function for
```classify_coverage_data: coverage_data -> coverage_gap option```,
which takes in `coverage_data` containing an expression and type and outputting an `option coverage_gap` containing `coverage_data` and the `reason`.

For this diff, we want to identify coverage gaps when the type is the plain `Any` type, and output the reason as `TypeIsAny`. This is because if an expression is of `Any` type, as `Any` is compatible with anything, bypassing checks even if the expression could crash at runtime, such as in the example `print(x + 1)`

### Stack diff:
In later diffs in this stack, we will address other variants such as
- `GenericParameterIsAny` where some parameter is `Any` (e.g., `List[Any]`)
- `CallableParameterIsAny` where the callable has some parameter is any
- `CallableParameterIsNonAnyNotUnknown` where argument expressions have non-`Any` types, but the Callable expects parameters that are `unknown`

Reviewed By: pradeep90

Differential Revision: D36953177

fbshipit-source-id: f63f4891012411503b6cab2990b0322abbc4dcef