Refactor typeCheck [1] split state-vs-compute for local_annotations
Summary:
I'm working on a refactor of typeCheck.ml that will allow me to move the
typeEnvironment state-handling logic into typeEnvironment.ml, which is
important in order to move fast and not make mistakes when implementing the
type environment overlay logic.

This commit splits `get_or_recompute_local_annotations` (whose name pretty
clearly indicates that it does two very different things) so that the
computation is a stand-alone function taking a global environment as input
and is separate from the part that checks a shared-memory table.

Reviewed By: pradeep90

Differential Revision: D36942243

fbshipit-source-id: f08287079445be5eb31e8410df5f3893dfa5a5c4