Refactor typeCheck [2] Split state-vs-compute for check
Summary:
This commit is part of a stack where I am decoupling the computational
part of `typeCheck.ml` (which should stay where it is) from the
portions that manage `TypeEnvironment` state (which need to move to
`typeEnvironment.ml` so that it's much easier to write and reason
about the overlay logic).

In this commit I split `run_on_define` so that the portion that
creates a global resolution from a read-only global environment and
actually invokes `check_function_definition` (including the handling
of the case where no define is found!) all lives in a pure function
`check_define`.

The `run_on_define` function - whose API is unchanged - now only
deals with state management: based on the config, do we save type
errors? Do we save local annotations?

Reviewed By: grievejia

Differential Revision: D36942241

fbshipit-source-id: d95d060e5710372883101224c38122ed716d69c3