Use a dependency tracked table for define names
Summary:
I want the function currently known as `all_defines_in_module`
to have dependency tracking - this is key for long-term maintainability
because it is a part of the type checking loop (and it is in fact
dependency tracked, but in a hacky way).

To that end, I'm moving the names out of `KeyTracker.DefineKeys` and
into a top-level `DefineNames` dependency-tracked table. The dependency
tracking isn't enabled yet as of this diff (it will take some time before
that is ready, since it requires a refactor of type checking) but this
code is much closer to where we want to be.

Per some team discussion, the choice of name `define_names` is also closer
to where we want to be in the future; I'm still using `all_defines_in_module`
for the API function because I don't want to change the interface in this
commit, but most of the internal code now refers to the references as
`define_names`.

Reviewed By: grievejia

Differential Revision: D36986744

fbshipit-source-id: 70edcd2a209a4a5757daee735838574c243a580a