Add test of (the function soon to be known as) `get_define_names`
Summary:
In my first attempt to refactor `all_defines_in_module` into a
dependency-tracked `get_define_names`, I broke the integ tests.

As a result, I though it was helpful to spin the tests out into a
separate commit so that we can verify the behavior is unchanged
(other than dependency tracking).

Reviewed By: grievejia

Differential Revision: D36991020

fbshipit-source-id: c0e9e7f98d0e11ae5081964acf309fcb2a3bfce0