Move get_define_body to ReadOnly module
Summary:
Since `get_define_body` is defined in terms of `get_function_definition`,
there's no reason to be defining it in the implementation of `read_only`;
that logic can live inside the ReadOnly module.

This will make it easier to create the overlay code since it's one less
callback we have to define.

Reviewed By: grievejia

Differential Revision: D36987324

fbshipit-source-id: e0dad3f7dac807c58e4fbbbd03d5eaf0e866345b