Move type environment state to typeEnvironment.ml
Summary:
This commit finally moves the existing functions from `typeCheck.ml`
that mutate a `TypeEnvironment.t` into `typeEnvironment.ml`.

The logic is completely unchanged, but I did rename them because the
names were kind of relying on the module name to make sense
(e.g. `TypeCheck.run_on_define` "runs type checking on a define", which
doesn't make as much sense when applied to `TypeEnvironment` as
`TypeEnvironment.populate_for_definition` which "populates the type environment
for a define").

All callsites are updated. One thing to note is that this flips the
order of module dependencies in the dune build so that `TypeEnvironment`
depends on `TypeCheck` instead of the other way around; this seems like
a good thing to me, the whole point of this refactor is that type check
really acts like the "produce value" logic for `TypeEnvironment` which
implies this is the correct conceptual ordering.

Reviewed By: pradeep90

Differential Revision: D36942245

fbshipit-source-id: e358881f9b366825a5c12f639e08087d4ee73546