Identify expressions with ContainerParameterIsAny coverage gaps
Summary:
### Context:
So far, we have only identified expressions with `TypeIsAny` coverage gaps. This diff identifies coverage gaps in the `ContainerParameterIsAny` variant.

### Problem:
We want to identify coverage gaps as they potentially cause runtime crashes. For the `ContainerParameterIsAny` we want to address types where the parameter is `Any`.

### Solution:
We extend the current `classify_coverage_data` function by adding cases in the pattern matching for these parametric types. This helps us handle cases such as `List[Any],Set[Any],Dict[Any,Any]`.

### Stack diff:
In this and previous diffs, we have addressed
- TypeIsAny where the type is the Any Type
- `ContainerParameterIsAny` where some parameter is `Any` (e.g., `List[Any]`)

In later diffs in this stack, we will address other variants such as
- `CallableParameterIsAny` where the callable parameters are all any
- `CallableParameterIsNonAnyNotUnknown` where argument expressions have non-`Any` types, but the Callable expects parameters that are `unknown`

Reviewed By: pradeep90

Differential Revision: D37009551

fbshipit-source-id: b158245bee020097e84cc49ba97fe2825f07febd