Fix a bug when a constructor is decorated with `functools.partial`
Summary:
There is a very specific edge case where we can have a constructor with `implicit_self = true` and `implicit_dunder_call = true`, when the constructor is decorated with a class decorator that uses `functools.partial`. See the integration test for an example.

This is a very complex case that we cannot handle precisely anyway since we call `functools.partial` which doesn't know which function is eventually called. Let's propagate the empty taint in that case as a quick workaround.

Reviewed By: dkgi

Differential Revision: D37021666

fbshipit-source-id: b9b357663c85b83824abadaa9a7b6e66e79e1d59