Address nit to use semicolons
Summary:
I'm actually not a fan of this change, but it was requested
as a nit in review so it's worth a debate.

In my opinion, ocaml side-effects are a *very* sharp edge
in our code because the type system is completely unaware of
them. As a result, I like them to be visually loud (log statements
are an exception to this general rule), and a `let () =` is
louder to my eye than a semicolon.

In addition, ocaml syntax errors are really nasty - some of the
hardest to eyeball of any language I've used. Semicolons make this
worse because they are a one-sided delimiter: notice how I have to
insert parentheses for this to work!

By contrast, `let () = ... in` creates
a 2-sided delimiter which makes syntax issues less likely, because
the parser can better guess our intent.

But if we think this is better we can ship it; I prefer abandoning
but my opinion is weakly held.

Reviewed By: pradeep90

Differential Revision: D36993279

fbshipit-source-id: edc8422784d4ee84b5d79899a12cb62e7dd99ba7