Handle hash collisions in dependencies
Summary:
In AstEnvironment, the only possible dependent of a RawSource
is a preprocessed source, registered via WildcardImport.

But because hashing isn't precise, it is possible for us to
have registered additional dependencies that use the same
hash. As a result, we need to ignore those; failing can cause
pyre to crash if there is a hash collision.

Reviewed By: grievejia

Differential Revision: D37039530

fbshipit-source-id: 953c8ec14765d6814f37f4a9e395d7de3b5e1cbd