Identify expressions with CallableParameterIsUnknownOrAny coverage gaps
Summary:
### Context:
So far, we have only identified expressions with `TypeIsAny` and `ContainerParameterIsAny` coverage gap variants. This diff identifies coverage gaps in the `CallableParameterIsUnknownOrAny` variant.

### Problem:
We want to identify coverage gaps as they potentially cause runtime crashes. For the `CallableParameterIsUnknownOrAny` we want to address callables where every parameter is `unknown` or `Any`. This catches the case where the function is unannotated, so all parameters would be `unknown`. It also catches when the parameters are annotated to be `Any`.

### Solution:
We extend the current `classify_coverage_data` function by adding cases in the pattern matching for these callable types. This helps us handle callables where every parameter is `unknown` or `Any`.

### Stack diff:
In this and previous diffs, we have addressed
- TypeIsAny where the type is the Any Type
- `ContainerParameterIsAny` where some parameter is `Any` (e.g., `List[Any]`)
- `CallableParameterIsUnknownOrAny` where the callable parameters are all any

### Limitations
In this diff, we only examined callables with ` Defined` and `Named` parameters. However, there are many other callables. In addition to `Defined` parameters, callable parameters can also be `Undefined` and `Variadic` types. In additional to `Named` parameters, they can be positional, keyword, variable, and keywords. In a later task, we will address these other coverage gaps in callables.

Reviewed By: pradeep90

Differential Revision: D37021551

fbshipit-source-id: c641dbe388b3b25d3bf449556777643e56cec759