ModuleTrackerTest, s/source_path/module_path/g
Summary:
In order to make code involving the BuildSystem type-safe,
I needed to rename the old `SourcePath` module to `ModulePath`.

This was a big win in terms of type safety and worthwhile, but
now there's a long tail of cleanups I need to do renaming variables
and functions.

I have a task to go do this, but for now I'll stick to doing it
a little bit at a time; right now the bad names in ModuleTrackerTest
is making it hard to write easy-to-follow overlay tests, so that's
what I'm going to prioritize.

Reviewed By: grievejia

Differential Revision: D37045916

fbshipit-source-id: d9082e0e2ef003c82240cf6a03ed671a22379aab