Get rid of UGE.UpdateResult.read_only
Summary:
It's a needless complication in our update result logic that we've
had to store a `read_only`; it was a hacky way of getting very limited
type safety in `update_only_this_environment` prior to the
`FromReadOnlyUpstream` refactor, but at this point there's no need.

If we want a read-only, we ought to get it from the read-write
environment instead of getting it out of an update result; i.e.
there should be only one obvious way to do it!

I already removed it from the functor used to create most layers, let's
strip it out of UGE.UpdateResult as well.

Reviewed By: grievejia

Differential Revision: D37055920

fbshipit-source-id: 08e735199aa46ebdda444c5428187faa7e220c6c