Treat index access to named tuples as attribute access in the forward analysis
Summary:
Index access to named tuples is the same as attribute access. For example,
```
>>> Point = namedtuple('Point', ['x', 'y'])
>>> p = Point(11, y=22)
>>> p.x
>>> p[0]
```
Line 3 and 4 are semantically the same.
Since this is current not modeled and may cause false negatives, this stack fixes this problem.
This diff fixes the forward analysis. The next diff will update the backward analysis.

Reviewed By: arthaud

Differential Revision: D36955450

fbshipit-source-id: 5ccc63369d61487f0c2bea50b563b99244983aa8