Add filesystem abstraction for ModuleTracker tests
Summary:
In code review, it was requested to use more descriptive names for
files in the overlay test.

That makes sense, but the existing tests are massive and it's probably
not worth going in and using more verbose names. So instead, this commit
creates a little DSL for describing test directory structure so that
we can inline that information while still factoring out the low-level
details into a helper function.

Reviewed By: grievejia

Differential Revision: D37071345

fbshipit-source-id: 265ec6f764348c4b72e97c5a474fac92a2049c7c