Modify coverage_data to have optional expression
Summary:
### Context:
In a previous task, we identified coverage gaps. With the ability to identify coverage gaps, we want to calculate %coverage for a module. To do this, we want the visitors to also accumulate expression.

### Problem:
However, one visitor stores type annotations, so there is no "expression" there to store. To combat this, we modify `coverage_data `to have an optional expression.

### Solution:
We changed the type of `coverage_data` to contain an optional expression. We updated the tests to reflect this and refactored to avoid multiple repetitive sections.

### Stack diff:
In this diff, we
- Modify `coverage_data` to contain an optional expression

In later diffs, we will
- Modify the two visitors to also accumulate expression
- Add a function to return list of tuples of `(Location.t, coverage_data)` instead of `(Location.t, Type.t)`
- Add a function to calculate %coverage using results from last task

Reviewed By: pradeep90

Differential Revision: D37070552

fbshipit-source-id: 34931151e482da971fcb9161cb105e7ebf7f3f15