Preserve binding mappings in filtered locals
Summary:
Keep the binding locations rather than only grabbing the keys of local bindings in the scope helper.

This is helpful both for having locations returned in find all references, but will also enable us to pull out the visitor logic that grabs access reads. We can diff the bindings with the accesses to avoid having to special-case statement matches and avoid grabbing assign targets, etc.

Reviewed By: grievejia

Differential Revision: D37025497

fbshipit-source-id: 76cb36c1b249b1afca92085b9abe5fe45ff5dea1