Add AstEnvironment.Overlay
Summary:
At the AstEnvironment layer, overlays only require one operation which is (as for module tracker) to update with new in-memory sources.

This needs to produce an UpdateResult that includes, as invalidated_modules, not only the directly invalidated sources but any dependencies of them via wildcard export.

These invalidated modules can be used by downstream updates.

Reviewed By: pradeep90

Differential Revision: D36992181

fbshipit-source-id: 480ca3c070b3f3a1e9c1a22499013776802765a7