Add UnannotatedGlobalEnvironment.Overlay (update only)
Summary:
UnannotatedGlobalEnvironment is more complicated than ModuleTracker and AstEnvironment because - due to the wildcard import dependencies coming from preprocessing - it is the first layer where we have to start worrying about dependency fan-out.

As a result its update function needs to update the underlying AstEnvironment overlay and also filter down the invalidated_modules to only those that are "owned" by the overlay - this prevents us from fanning out unsaved changes to downstream wildcard imports.

For the same reason, this layer will eventually need a function that propagates parent updates, but that is out of scope for this diff: I want to get the basic functionality for an overlay (how does it behave given a static parent) in place before worrying about parent propagation.

The implementation of `read_only` here is nontrivial because there are so many data structures owned by `UnannotatedGlobalEnvironment`. The strategy is:
- for all functions of the form `all_.*`, we just use the parent environment; these functions are not used except by `pyre query`, debugging tools in `check`, and pysa - none of which should ever need overlay support
- for all other functions, we follow a standard pattern:
   - first do an ownership check for the key
   - if the key is not owned, proxy to the corresponding function on the parent
   - otherwise, use the read-only backed by our own `from_upstream_readonly` (which should contain only the overlay data, a "skinny" dataset)

Reviewed By: pradeep90

Differential Revision: D36992180

fbshipit-source-id: 63db0d82f6e317ee929ecca3558d994d5e102dce