Add ModuleTracker.Overlay
Summary:
At the module tracker level, overlays only really need an update operation which adds a new or updated in-memory source.

These operations always produce NewExplicit updates to consume downstream.

A complication is that there are many ModulePaths that may pertain to the same tracked qualifier - for example, stubs will shadow code and `local_root` files will shadow external files.

The invariants for how we do this are:
- if an ArtifactPath from an unsaved change can be mapped to a ModulePath, then we
  will store the unsaved code and mark that qualifier as owned by the overlay. So the overlay
  *always* contains the latest unsaved code from tracked modules.
- we only override the `get_raw_source` function; all other functions (i.e. everything that
  handles mapping qualifiers to ModulePaths) comes from the parent.

A consequence of this is that if there is some file shadowing an overlaid file, the overlaid code
won't actually appear in any downstream environment - the data will be there in ModuleTracker.Overlay,
but `get_raw_source` will always be passed the higher-priority `ModulePath`.

The reason we want to store overlay code even if it is shadowed is twofold:
- this makes the invariants much simpler, so we can reason about correctness.
  That's the big reason - trying to think through all the state transitions is probably a
  bad way of architecting overlays, since there will be too many possible transitions.
- but if we do want to think about concrete edge cases, consider the case where we
   open a module `code.py` that has a stub, make some changes, and then the stub gets
   deleted on disk. If we had failed to save those unsaved changes at the appropriate ModulePath
   (which at the time was shadowed), then we won't be able to start providing editor functionality
   once the stub is removed.

Reviewed By: pradeep90

Differential Revision: D36992179

fbshipit-source-id: 4bc4d6b3d78804d479251cb6ad057b0c13359ab3