Use a try/except to degrade gracefully in 3.7
Summary:
I had hoped that we could bypass failures in the `find_symbols` module by
dropping support for Python 3.6, which is past end-of-life, but it turns
out the API we're using doesn't exist in 3.7 either and that's still a year
away from end-of-life.

So, for now we should catch the resulting AttributeError and degrade
gracefully.

Differential Revision: D37091905

fbshipit-source-id: a1981b9aec88fa77270c7c910e3a3f649c4b6577