find_symbols logic should degrade gracefully in 3.7 (#619)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/619

I had hoped that we could bypass failures in the `find_symbols` module by
dropping support for Python 3.6, which is past end-of-life, but it turns
out the API we're using doesn't exist in 3.7 either and that's still a year
away from end-of-life.

So, for now we should catch the resulting AttributeError and degrade
gracefully. There's a test in `persistent_test` that proves this degredation
works; in the interests of readability I just skip the find_symbols test entirely
in 3.7 since we know almost every function returns degraded results.

Reviewed By: grievejia

Differential Revision: D37091905

fbshipit-source-id: 6143ba1a38a8eef5571619be55749080bc7e796b