Clean up environment [3a] generalize the update logic
Summary:
In order to implement the EnvironmentTable.Overlay module,
I'm going to need the ability to customize the list of
dependency triggers that we include in `update_only_this_environment`
becuase I need ownership filtering, which needs to live in the
Overlay module rather than the FromReadOnlyUpstream module.

As a result, I want the `update_only_this_environment` to take
a filtereds list of triggered dependencies as input, which will
give me the ability to customize the filtering in the Overlay
update functions.

This commit is a first step toward doing this: it updates
the interface to take a list of triggered dependencies as input,
which is the type signature we want.

In the next commit I'll extract the filtering logic.

Reviewed By: grievejia

Differential Revision: D37088845

fbshipit-source-id: 6c0e517a3343e6cfcd7710d6455499a66e383910