Clean up environment [1] use PreviousEnvironment.S
Summary:
In the `environment.ml` module, it's easy to get a little turned around by the
fact that we use puns of module type names versus the names of module inputs to
functors.

Since we already have `UpdateResult.S` versus `UpdateResult` to keep this
clear, we can just use the same pattern for `PreviousEnvironment` to make this
clearer.

Reviewed By: grievejia

Differential Revision: D37081907

fbshipit-source-id: 559154b7115d441523f0ed042bf7dd37e927aacb