Clean up environment [2] Deduplicate some code
Summary:
In this commit, I pick up all of the signature of `PreviousEnvironment.S`
and include it in `Environment.S`.

The advantages of this are:
- less duplication - now I only need to update two signatures if I change
  something, and it's quicker to scroll through the file (which can be
  confusing, especially with so many similar-looking code blocks)
- it is now clearer that the output type of the functor does in fact
  meet the `PreviousEnvironment` interface requirement; before this was
  not self-evident, and in fact it was easy to accidentally add a bug
  where it wasn't true and get low-signal error messages

Note that we can't do the same trick to eliminate duplication in
`EnvironmentTable.S` - this is because the `ReadOnly` type in
`EnvironmentTable.S` needs to be more specific.

Reviewed By: grievejia

Differential Revision: D37088844

fbshipit-source-id: 892d0cce5b4a16f7968224b1d2208fefc104e36d