Clean up source_path name: rename functions in ModuleTracker
Summary:
About a month back, we decided to rename `SourcePath` to `ModulePath`
because (a) the name is more descriptive and (b) the term was overloaded,
since we also used "source path" to mean some path in the directories
specified under `Configuration.Analysis.source_paths` - i.e. pre-build-system
code paths.

Today, only the second meaning is valid but we still have many functions and
variables that are misnamed. In this commit, I rename all the module tracker
functions.

Reviewed By: grievejia

Differential Revision: D37109797

fbshipit-source-id: b123c28b5aeeeb685714615f58ee4421ce2f15d9