Handle integer constants that are greater than OCaml INT_MAX
Summary:
Currently integer constants that are greater than OCaml INT_MAX are treated as a normal Constant.integer with capped value (Int.max_value). To properly handle it, we need a separate Constant type for big integer and propagate it all the way up in AST.

As a bootcamp task, this diff only create a separate Constant.BigInteger type for it and correct the revealed type for code as in the test plan.

Reviewed By: grievejia

Differential Revision: D37124022

fbshipit-source-id: 343840f0eda0a057f96317b2fb781f5222d5c2ac