avoid importing `typing.TypedDict`
Summary:
We export `pyre_extensions.tests` in the `__init__.py`,
which is surprising but also intended. So any lack of
backward compatibility with python will break not only
the tests but the entire extensions module even in production.

Now that we test pyre_extensions in github CI, we should be
able to catch such problems prior to release in the future.

Differential Revision: D37178251

fbshipit-source-id: b0f7f42a25a0dfacfdc0817dabb788a26a63b3f2