avoid importing `typing.TypedDict` (#630)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/630

We export `pyre_extensions.tests` in the `__init__.py`,
which is surprising but also intended. So any lack of
backward compatibility with python will break not only
the tests but the entire extensions module even in production.

Now that we test pyre_extensions in github CI, we should be
able to catch such problems prior to release in the future.

Reviewed By: grievejia

Differential Revision: D37178251

fbshipit-source-id: 0d6f2c0c7064c200e2780fd1ab10016b2930327b