Test pyre_extensions in github CI (#629)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/629

We accidentally broke 3.7 support by importing `typing.TypedDict`.

The only way to catch this reliably is to test the unreleased code
on github CI, which we currently don't do because only `client/`
tests are run.

As part of doing this, I also rename `simple_tests.py`, which is violating
our test module naming convention, to `basic_test.py`.

Reviewed By: grievejia

Differential Revision: D37177511

fbshipit-source-id: 68111d5c7f22a5b05c5c89eb72d7fa755156b850