Handle non-unicode binary data in the client
Summary:
The ocaml backend of pyre works with bytestrings. This
in itself isn't a problem - there's nothing inherently
wrong with using bytestrings for string-or-binary data,
in fact that's what shiny new parser libraries like
tree-sitter do.

But we have to be careful when we send that data over the
wire; Python wants to consume it as json, whcih has to be
valid unicode (which we'll try to get using the default
encoding utf-8).

For the vast majority of python files, variable names and
strings are utf-8 so naively sending the ocaml bytestrings
over works fine.

But there are a few edge cases where things go wrong:
- for certain, a bytes literal in python - even a utf-8 encoded
  module - will parse into an ocaml byte string that is not
  valid. This is because when CPython parses a bytes literal,
  the resulting AST node does not contain the string representing
  those bytes, it contains the bytes themselves (which may not be
  a valid encoding)
- it's possible that a unicode string literal in an (e.g.) latin-1
  module might wind up being latin-1 encoded in ocaml. I'm
  actually unsure whether this happens; CPython's parser very
  likely will utf-8 encode for us.

Differential Revision: D37211803

fbshipit-source-id: 409d86ab677a4bc20d3c384fe9baf32a76fa4fac