Get coverage gaps in a module
Summary:
### Context:
In a previous task, we identified coverage gaps. With the ability to identify coverage gaps, we want to calculate %coverage for a module.

### Problem:
To do this, we want to get the coverage gaps for all the expressions in a module. Then, we would be able to quantify the percent of coverage by looking at the total number of expressions and number of coverage gaps.

### Solution:
As the lookup input is a `coverage_data Location.Table.t`, we classify each `(Location.t * coverage_data)` into a coverage gap using `classify_coverage_data`. Then, we use `List.filter_map` to get a list of expressions that have `Some` coverage gap.

### Stack diff:
In earlier diffs, we
- Modify `coverage_data` to contain an optional expression
- Modify the two visitors to also accumulate expression
- Add a function to return list of tuples of `(Location.t * coverage_data)` instead of `(Location.t * Type.t)`

In this diff, we
- Add a function to calculate get coverage gaps using results from last task

Reviewed By: pradeep90

Differential Revision: D37124365

fbshipit-source-id: 48f7566078273c3db274f5ce990b329666bbeec1