Add variants and types for expression level coverage
Summary:
### Context:
In previous tasks, we made functions to get coverage gaps for a module. In order to expose this functionality to the user, we want to add a query called `expression_level_coverage(paths)`.

### Problem:
For our query, we want to return a response for a list of paths, containing the path name, total number of expressions, and coverage gaps for each path. With this, we can quantify coverage using the number of coverage gaps and number of total expressions and find which locations have coverage gaps.

In this diff, we want to set up request and response variants for this query, as well as types that would be useful for implementing this variant.

### Solution:
We add `ExpressionLevelCoverage` as a request variant to accept a list of paths.

We add `ExpressionLevelCoverageResponse` as a response variant to return a list of paths, total number of expressions, and `coverage_gaps` for each path, which we represent with a `coverage_at_path` type.

We represent `coverage_gaps` by a list of `coverage_gap_by_location` type, which includes `Location.t,Type.t,reason` but not `Expression.t` for brevity.

### Stack diff:
In this diff, we
- Add response and request variants as well as other useful types

In later diffs, we will
- Implement the actual response to the query

Reviewed By: pradeep90

Differential Revision: D37192199

fbshipit-source-id: cc2230285bf26c1ff5e7307bd84cb963464b77e6