Modify visitors to also accumulate expression
Summary:
### Context:
In a previous task, we identified coverage gaps. With the ability to identify coverage gaps, we want to calculate %coverage for a module.

### Problem:
To do this, we want the visitors to also accumulate expression.

### Solution:
We changed `resolved_type_lookup` from a `Type.t Location.Table.t` into `coverage_data Location.Table.t`.

### Stack diff:
In this diff and earlier diffs, we
- Modify `coverage_data` to contain an optional expression
- Modify the two visitors to also accumulate expression
- Add a function to return list of tuples of `(Location.t, coverage_data)` instead of `(Location.t, Type.t)`

In later diffs, we will
- Add a function to calculate get coverage gaps using results from last task

Reviewed By: pradeep90

Differential Revision: D37054002

fbshipit-source-id: 82b86e830215b767d1267d5179cf0ab480f0a649